fixed PisoItemListView
Problem that has been solved: The selection of items was based on the name of the boxes, and not on the box objects themselves. Thus, as some of the pisos have boxes with the same name, the filtering was not done correctly, and some items appeared in the lists of the wrong pisos.

Solution: Base the filtering on the box objects and not on box names.